feat: add optional deadline to device calls, allow specifying of node id (#11)
* feat: allow specification of node id when adding node to config

* feat: add optional deadline to device calls

* chore: update synapse-api

* fix: switch node_id fields to uint32_t to match protos

* chore: make deadline truly optional with default nullopt

* chore: rename from deadline to timeout

* chore: update libndtp

+++ include/science/synapse/node.h
@@ -22,12 +22,12 @@ class Node {
   explicit Node(const synapse::NodeType& node_type);
   virtual ~Node() = default;
 
-  [[nodiscard]] auto id() const -> uint64_t;
+  [[nodiscard]] auto id() const -> uint32_t;
   [[nodiscard]] auto set_device(const Device* device) -> science::Status;
   auto to_proto(synapse::NodeConfig* proto) -> void;
 
  protected:
-  uint64_t id_;
+  uint32_t id_;
   synapse::NodeType type_;
   const Device* device_;
